Amended button to see currently chosen documents to fit in slightly different architecture
