+++ src/choose_documents.py
@@ -1,4 +1,5 @@
 import os
+import subprocess
 from PySide6.QtWidgets import QApplication, QFileDialog
 
 def choose_documents_directory():
@@ -15,6 +16,18 @@ def choose_documents_directory():
         for file_path in file_paths:
             symlink_target = os.path.join(docs_folder, os.path.basename(file_path))
             os.symlink(file_path, symlink_target)
+
+
+def see_documents_directory():
+    current_dir = os.path.dirname(os.path.realpath(__file__))
+    docs_folder = os.path.join(current_dir, "Docs_for_DB")
+
+    # Ensure the directory exists
+    if not os.path.exists(docs_folder):
+        os.mkdir(docs_folder)
+
+    # Open the directory in Windows File Explorer
+    subprocess.Popen(f'explorer "{docs_folder}"')
 
 if __name__ == '__main__':
     app = QApplication([])
